requirements.txt
protobuf~=5.29.1
python-dotenv~=1.0.1
torchmetrics~=1.6.0
levenshtein~=0.26.1